[userdata] Add userdata command to execute EC2 userdata with IMDSv2
This commit introduces a new command which fetches and executes AWS EC2 userdata using IMDSv2. Before this commit, iPXE was not compatible with IMDSv2 which requires a token in the header to fetch the user data. To fetch the token, a PUT request needs to be sent to IMDS. The token is then used in the header of the subsequent GET request to fetch the userdata. This required modifications to the httpcore.c file to add support for PUT requests and for the necessary headers.