[json-params] Add special handling for JSON parameters in http_form_params
Previously, all parameters were encoded as form data (key=value&key2=value2
format with URI encoding). This doesn't work for JSON payloads which need
to be sent as raw JSON strings.

Added check for PARAMETER_JSON flag:
- Form parameters: encoded as "key1=value1&key2=value2" with URI encoding
- JSON parameters: return just the key as-is (key contains the full JSON)

This is a temporary solution - JSON handling should be moved to a separate
function in the future.